[iOS] Suppress page count recalc when setting a snapshot paper rect https://bugs.webkit.org/show_bug.cgi?id=170156 <rdar://problem/30787872>
Reviewed by Tim Horton.

When _WKWebViewPrintFormatter computes a page count for snapshotting, it updates the paper
rect, which triggers UIKit to ask for the page count again while we're in the middle of
printing, putting us in a bad state.

Suppress page count recalc while setting the new paper rect, since we know this will not
cause it to change.

* UIProcess/_WKWebViewPrintFormatter.mm:
(-[_WKWebViewPrintFormatter _setSnapshotPaperRect:]): Temporarily set
_suppressPageCountRecalc to YES before updating the paper rect.
(-[_WKWebViewPrintFormatter _setNeedsRecalc]): Only call [super _setNeedsRecalc] if
_suppressPageCountRecalc is NO.
* UIProcess/_WKWebViewPrintFormatterInternal.h: Declared -_setNeedsRecalc.

Identifier: 187074@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@214449 268f45cc-cd09-0410-ab3c-d52691b4dbfc