Ensure that Node::willBeDeletedFrom() always removes touch event handlers from the document https://bugs.webkit.org/show_bug.cgi?id=170323 rdar://problem/23647630
Reviewed by Chris Dumez.

There are two instances where nodes are registered as touch event handlers without
having normal touch event listeners: slider thumb elements, and elements with overflow scrolling,
on iOS.

For such nodes, hasEventTargetData() will be false, but we want to ensure
that they are removed from the Document's touchEventHandler set, so move the
call to document.removeTouchEventHandler() outside of the conditional block.

This should be cheap in most cases when the touchEventHandler is empty.

* dom/Node.cpp:
(WebCore::Node::willBeDeletedFrom):

Identifier: 187242@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@214649 268f45cc-cd09-0410-ab3c-d52691b4dbfc