[WK2] Tapping editable text inside of a range selection no longer changes the selection to a caret https://bugs.webkit.org/show_bug.cgi?id=170327 <rdar://problem/31363816>
Reviewed by Tim Horton.

Source/WebKit2:

Currently, we're forcing all text interaction gestures to duck in lieu of data interaction gestures
when we should only be doing so for gestures that begin a loupe. This prevents other gestures, such as
single taps, from changing the selection when they should be allowed to.

Hooks into new UIKit SPI to make this tweak.
Introduces a new LayoutTest: editing/selection/caret-after-tap-in-editable-selection.html.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView pointIsInAssistedNode:]):
(-[WKContentView textInteractionGesture:shouldBeginAtPoint:]):

LayoutTests:

Adds a new WK2 interaction test for iOS verifying that tapping a selection in editable content sets the selection
to a caret, rather than maintaining the range selection. This test is disabled in OpenSource, since it relies on
synthetic touch events.

* TestExpectations:
* editing/selection/caret-after-tap-in-editable-selection-expected.txt: Added.
* editing/selection/caret-after-tap-in-editable-selection.html: Added.


Identifier: 187269@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@214687 268f45cc-cd09-0410-ab3c-d52691b4dbfc