B3 -O1 should generate better code than -O0 https://bugs.webkit.org/show_bug.cgi?id=170563
Reviewed by Michael Saboff.

Prior to this change, code generated by -O1 ran slower than code generated by -O0. This turned
out to be because of reduceStrength optimizations that increase live ranges and create register
pressure, which then creates problems for linear scan.

It seemed obvious that canonicalizations that help isel, constant folding, and one-for-one
strength reductions should stay. It also seemed obvious that SSA and CFG simplification are fast
and harmless. So, I focused on removing:

- CSE, which increases live ranges. This is a risky optimization when we know that we've chosen
  to use a bad register allocator.

- Sophisticated strength reductions that create more code, like the insane division optimization.

- Anything that inserts basic blocks.

CSE appeared to be the cause of half of the throughput regression of -O1 but none of the compile
time. This change also reduces the running time of reduceStrength by making it not a fixpoint at
optLevel<2.

This makes wasm -O1 compile 17% faster. This makes wasm -O1 run 19% faster. This makes -O1 code
run 3% faster than -O0, and compile about 4% slower than -O0. We may yet end up choosing to use
-O0, but at least now -O1 isn't totally useless.

* b3/B3ReduceStrength.cpp:



Identifier: 187486@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215057 268f45cc-cd09-0410-ab3c-d52691b4dbfc