Have parseRevisionProperty use default repo name as the Buildbot key. https://bugs.webkit.org/show_bug.cgi?id=170696
Patch by Kocsen Chung <kocsen_chung@apple.com> on 2017-04-10
Reviewed by Daniel Bates.

Previously, we needed to set a key and a fallbackKey when addressing
Buildbot data for Internal. That is no longer necessary as we can make a safe assumption
that the repo name maps exactly to that key for every repository other
than WebKit.

* BuildSlaveSupport/build.webkit.org-config/public_html/dashboard/Scripts/BuildbotIteration.js:
Set `key = repositoryName` and only use fallback keys when dealing with WebKit.

Identifier: 187641@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215215 268f45cc-cd09-0410-ab3c-d52691b4dbfc