test262: test262/test/annexB/language/comments/multi-line-html-close.js https://bugs.webkit.org/show_bug.cgi?id=170648
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2017-04-11
Reviewed by Keith Miller.

JSTests:

* test262.yaml:

Source/JavaScriptCore:

* parser/Lexer.cpp:
(JSC::Lexer<T>::lex):
A multi-line comment that contains a line terminator is itself treated
like a line terminator. An HTML Close Comment that comes after it can
therefore treat it like it is at the start of a line, because it was
immediately preceeded by the equivalent of a line terminator.

LayoutTests:

* js/parser-xml-close-comment-expected.txt:
* js/script-tests/parser-xml-close-comment.js:
Update test coverage for JavaScript HTML comments.

Identifier: 187661@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@215235 268f45cc-cd09-0410-ab3c-d52691b4dbfc